新版app/requirements.txt
# =============================================================================
# requirements.txt - JK-BMS Gateway V2.2.1
# =============================================================================
pyserial==3.5
paho-mqtt==1.6.1
PyYAML==6.0.1
